feat: :recycle: pagelink conversion to front matter function
pagelink was converted from a shortcode into a front matter function called link to simplify link creation. It works relatively the same as before, though collections does not need to be passed to it. Also, it shortens the actual call and is easier to understand at first glance.

pagelink will no longer work. Refactor to use the link front matter function